Stage 1: Points calculation and verification of Monte Carlo Method
This stage of development introduced the points calculation mechanism- using player expected goals and assists as a basis of a Poisson distribution calculation. Applied this to create a function calculating weekly points return using FPL game rules. Verified the convergence of the Monte Carlo method and the program to analytical solutions.