Update Stage 2: Clean up by adding functions
Put the captain selector into a function rather than the main script so it can easily be distinguished and called. Unified the points calculator into one function and introduced a new "position" variable to select which formula to use. Intention is for these functions to eventually extract specific elements of a dataframe to get the desired outputs.